Reparative snapshotting
As snapshots can be invalidated we can end up with entities that have
hundreds of events and no snapshot. The previous implementation relied
on the modulus of the version and the snapshot threshold being checked.

This meant that for a snapshot threshold of 10 and an existing entity
on version 251 in the event of the snapshot being invalidated for
whatever reason, we would replay all of the events every time we
interact with the entity until it reached version 260.

This change now also checks the difference between the first version set
and the current version in order to determine whether the entity should
be snapshotted so the snapshots will be aggressively taken in a manner
similar to how they are discarded.

It also moves the logic of whether a snapshot should be taken to a
method on the entity to allow clients of the gem to override the
behaviour however they deem necessary.